# src/main/gen/api/openapi.yaml
openapi: 3.0.1
info:
  description: "API de gestión de clínicas médicas, maquinaria, técnicos, tickets\
    \ y repuestos (AA1)"
  title: aa1
  version: 1.0.0
servers:
- url: http://localhost:8080
paths:
  /clinics:
    get:
      description: "Devuelve una lista con todas las clínicas registradas. \nMuestra\
        \ la información básica a través de **ClinicOutDto**.\n"
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ClinicOutDto'
                type: array
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
          description: Internal server error
      summary: Devuelve la lista de clínicas
      x-accepts:
      - application/json
    post:
      description: Registra una nueva clínica validando que el nombre no esté vacío.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Clinic'
        description: Datos de la nueva clínica (id no es necesario enviarlo)
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Clinic'
          description: Clínica registrada con éxito
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
          description: Bad request - Error de validación (NotBlank)
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
          description: Internal server error
      summary: Registrar una clínica
      x-content-type: application/json
      x-accepts:
      - application/json
  /clinics/{id}:
    delete:
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "204":
          description: Clínica eliminada correctamente (No Content)
      summary: Eliminar una clínica
      x-accepts:
      - application/json
    get:
      parameters:
      - description: ID único de la clínica
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClinicDto'
          description: Clínica encontrada
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
          description: Not Found - La clínica no existe
      summary: Buscar clínica por ID
      x-accepts:
      - application/json
    put:
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Clinic'
        required: true
      responses:
        "200":
          description: Clínica actualizada correctamente
        "404":
          description: No encontrada
      summary: Modificar una clínica
      x-content-type: application/json
      x-accepts:
      - application/json
  /machines:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Machine'
                type: array
          description: OK
      summary: Devuelve la lista de máquinas
      x-accepts:
      - application/json
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Machine'
        required: true
      responses:
        "201":
          description: Máquina registrada
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
          description: Bad request
      summary: Registrar una máquina
      x-content-type: application/json
      x-accepts:
      - application/json
  /machines/{id}:
    delete:
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "204":
          description: Máquina eliminada
      summary: Eliminar una máquina
      x-accepts:
      - application/json
    get:
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Machine'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
          description: Máquina no encontrada
      summary: Buscar máquina por ID
      x-accepts:
      - application/json
    put:
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Machine'
        required: true
      responses:
        "200":
          description: Máquina actualizada
      summary: Modificar una máquina
      x-content-type: application/json
      x-accepts:
      - application/json
  /technicians:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Technician'
                type: array
          description: OK
      summary: Devuelve la lista de técnicos
      x-accepts:
      - application/json
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Technician'
        required: true
      responses:
        "201":
          description: Técnico creado
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
          description: Error de validación
      summary: Registrar un técnico
      x-content-type: application/json
      x-accepts:
      - application/json
  /technicians/{id}:
    delete:
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "204":
          description: Técnico eliminado
      summary: Eliminar un técnico
      x-accepts:
      - application/json
    get:
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Technician'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
          description: Técnico no encontrado
      summary: Buscar técnico por ID
      x-accepts:
      - application/json
    put:
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Technician'
        required: true
      responses:
        "200":
          description: Técnico actualizado
      summary: Modificar un técnico
      x-content-type: application/json
      x-accepts:
      - application/json
  /tickets:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Ticket'
                type: array
          description: OK
      summary: Devuelve la lista de tickets
      x-accepts:
      - application/json
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Ticket'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ticket'
          description: Ticket creado con éxito
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
          description: Error en los datos obligatorios
      summary: Registrar un ticket de avería
      x-content-type: application/json
      x-accepts:
      - application/json
  /tickets/{id}:
    delete:
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "204":
          description: Ticket eliminado
      summary: Eliminar un ticket
      x-accepts:
      - application/json
    get:
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ticket'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
          description: Ticket no encontrado
      summary: Buscar ticket por ID
      x-accepts:
      - application/json
    put:
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Ticket'
        required: true
      responses:
        "200":
          description: Ticket actualizado
      summary: Modificar un ticket
      x-content-type: application/json
      x-accepts:
      - application/json
  /spare-parts:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/SparePart'
                type: array
          description: OK
      summary: Devuelve la lista de repuestos
      x-accepts:
      - application/json
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SparePart'
        required: true
      responses:
        "201":
          description: Repuesto creado
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
          description: Error de validación (BadRequest)
      summary: Registrar un nuevo repuesto
      x-content-type: application/json
      x-accepts:
      - application/json
  /spare-parts/{id}:
    delete:
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "204":
          description: Repuesto eliminado
      summary: Eliminar un repuesto
      x-accepts:
      - application/json
    get:
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SparePart'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
          description: Repuesto no encontrado
      summary: Buscar repuesto por ID
      x-accepts:
      - application/json
    put:
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SparePart'
        required: true
      responses:
        "200":
          description: Repuesto actualizado
      summary: Modificar un repuesto
      x-content-type: application/json
      x-accepts:
      - application/json
components:
  schemas:
    Clinic:
      example:
        id: 1
        name: Clínica San José
      properties:
        id:
          format: int64
          type: integer
        name:
          type: string
        address:
          type: string
        phone:
          type: string
        email:
          type: string
      required:
      - name
      type: object
    Machine:
      example:
        id: 10
        serialNumber: SN-777-B
        model: X-Ray v2
        manufacturer: General Electric
      properties:
        id:
          format: int64
          type: integer
        serialNumber:
          type: string
        model:
          type: string
        manufacturer:
          type: string
        clinic:
          $ref: '#/components/schemas/Clinic'
      required:
      - manufacturer
      - model
      - serialNumber
      type: object
    Technician:
      example:
        id: 5
        name: Alberto
        surname: García
        email: alberto.tec@clinic.com
      properties:
        id:
          format: int64
          type: integer
        name:
          type: string
        surname:
          type: string
        email:
          format: email
          type: string
        phone:
          type: string
        active:
          default: true
          type: boolean
      required:
      - email
      - name
      - surname
      type: object
    Ticket:
      example:
        id: 100
        description: Fallo en el equipo
        openDate: 2025-12-28
        status: OPEN
      properties:
        id:
          format: int64
          type: integer
        description:
          type: string
        openDate:
          format: date
          type: string
        closeDate:
          format: date
          type: string
        status:
          type: string
        machine:
          $ref: '#/components/schemas/Machine'
        technician:
          $ref: '#/components/schemas/Technician'
      required:
      - description
      - openDate
      - status
      type: object
    SparePart:
      example:
        id: 20
        name: Filtro de aire
        serialNumber: SP-55-AA
        manufacturer: MedicalParts
        price: 45.5
        stock: 10
      properties:
        id:
          format: int64
          type: integer
        name:
          type: string
        serialNumber:
          type: string
        manufacturer:
          type: string
        price:
          format: double
          type: number
        stock:
          format: int32
          type: integer
      required:
      - manufacturer
      - name
      - price
      - serialNumber
      - stock
      type: object
    ClinicDto:
      example:
        address: address
        name: name
        id: 0
      properties:
        id:
          format: int64
          type: integer
        name:
          type: string
        address:
          type: string
      type: object
    ClinicOutDto:
      example:
        name: name
        id: 0
      properties:
        id:
          format: int64
          type: integer
        name:
          type: string
      type: object
    ErrorResponse:
      properties:
        code:
          type: integer
        message:
          type: string
        title:
          type: string
      type: object
    InternalServerError:
      allOf:
      - $ref: '#/components/schemas/ErrorResponse'
      example:
        code: 500
        message: Error interno
        title: internal-server-error
    BadRequest:
      allOf:
      - $ref: '#/components/schemas/ErrorResponse'
      example:
        code: 400
        message: Validation error
        title: bad-request
    NotFound:
      allOf:
      - $ref: '#/components/schemas/ErrorResponse'
      example:
        code: 404
        message: Resource not found
        title: not-found