# src/main/java/com/example/aa1/dto/ClinicOutDto.java
package com.example.aa1.dto;

public class ClinicOutDto {

    private Long id;
    private String name;

    public Long getId() {
        return id;
    }

    public void setId(Long id) {
        this.id = id;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }
}